refactor(app): split tauri command modules
